refactor: update button component to use new variants and loading props
- Changed button styles to utilize button.variants instead of button.styles
- Integrated Loading component for spinner functionality with customizable loading color and variant
- Updated ButtonProps to include optional loadingColor and loadingVariant properties
