+++ src/components/ui/button/button.types.ts
@@ -1,7 +1,13 @@
 import { ReactNode } from "react";
-import { ButtonVariants } from "./button.styles";
+import { ButtonVariants } from "./button.variants";
+import { LoadingVariants } from "../loading/loading.variants";
+
+type LoadingColor = LoadingVariants["color"];
+type LoadingVariant = LoadingVariants["variant"];
 
 export type ButtonProps = {
   children: ReactNode;
+  loadingColor?: LoadingColor;
+  loadingVariant?: LoadingVariant;
 } & ButtonVariants &
   React.ButtonHTMLAttributes<HTMLButtonElement>;
